Extract common table implementation
